+++ packages/stacked/example/lib/app/app.logger.dart
@@ -62,8 +62,8 @@ class SimpleLogPrinter extends LogPrinter {
       var currentStack = StackTrace.current;
       var formattedStacktrace = _formatStackTrace(currentStack, 3);
 
-      var realFirstLine =
-          formattedStacktrace?.firstWhere((line) => line.contains(className));
+      var realFirstLine = formattedStacktrace
+          ?.firstWhere((line) => line.contains(className), orElse: () => "");
 
       var methodName = realFirstLine?.replaceAll('$className.', '');
       return methodName;
@@ -136,7 +136,7 @@ Logger getLogger(
       exludeLogsFromClasses: exludeLogsFromClasses,
     ),
     output: MultipleLoggerOutput([
-      ConsoleOutput(),
+      if (!kReleaseMode) ConsoleOutput(),
     ]),
   );
 }