+++ packages/stacked/lib/src/code_generation/stacked_logger_annotations.dart
@@ -7,8 +7,13 @@ class StackedLogger {
   /// loggerOutputs must extends [LogOutput]
   final List<Type> loggerOutputs;
 
+  /// When set to true console logs will not be printed in release mode
+  /// Default is true
+  final bool disableReleaseConsoleOutput;
+
   const StackedLogger({
     this.loggerOutputs = const [],
     this.logHelperName = 'getLogger',
+    this.disableReleaseConsoleOutput = true,
   });
 }