+++ packages/stacked_generator/lib/src/generators/logging/logger_class_content.dart
@@ -1,6 +1,7 @@
 const String LogHelperNameKey = 'logHelperName';
 const String MultiLoggerImports = 'MultiLoggerImport';
 const String MultipleLoggerOutput = 'MultiLoggerList';
+const String DisableConsoleOutputInRelease = 'MultiLoggerList';
 
 const String loggerClassContent = '''
 // ignore_for_file: avoid_print
@@ -140,6 +141,7 @@ Logger $LogHelperNameKey(
       exludeLogsFromClasses: exludeLogsFromClasses,
     ),
     output: MultipleLoggerOutput([
+      $DisableConsoleOutputInRelease
       ConsoleOutput(),
       $MultipleLoggerOutput
     ]),