+++ packages/stacked_generator/lib/src/generators/logging/logger_config.dart
@@ -4,6 +4,10 @@ class LoggerConfig {
   final String logHelperName;
   final Set<String> imports;
   final List<String> loggerOutputs;
+
+  /// When set to true console logs will not be printed in release mode
+  /// Default is true
+  final bool disableReleaseConsoleOutput;
   // Future
   // final bool enableGoogleCloudLogging;
 
@@ -11,5 +15,6 @@ class LoggerConfig {
     this.imports = const {},
     this.loggerOutputs = const [],
     this.logHelperName = 'getLogger',
+    this.disableReleaseConsoleOutput = true,
   });
 }