+++ packages/stacked_services/lib/src/dialog/dialog_service.dart
@@ -77,7 +77,9 @@ class DialogService {
         title: title,
         description: description,
         cancelTitle: cancelTitle,
+        cancelTitleColor: cancelTitleColor,
         buttonTitle: buttonTitle,
+        buttonTitleColor: buttonTitleColor,
         dialogPlatform: dialogPlatform,
         barrierDismissible: barrierDismissible,
       );