+++ packages/stacked/CHANGELOG.md
@@ -1,8 +1,37 @@
+## 2.3.0
+
+### New Feature: Adds per field validation
+
+- A setValidationMessage will now be generated for each field for you to set and use similar to the global validation message that was used before
+
+- Breaking changes on `FormViewModel`
+  - `showValidation` → `showValidationMessage`
+
+## 2.2.8
+
+- Adds `disableReleaseConsoleOutput` to the logger annotation to disable console output in release mode
+
+## 2.2.7+1
+
+- Upgraded the dependency on `provider` to 6.0.0
+
+## 2.2.7
+
+- Fixes non-reactive `ViewModelBuilderWidget`
+
+## 2.2.6
+
+- Optimized skeleton loader to be dynamic
+
 ## 2.2.5
+
 - Added IgnorePointer to Skeleton Loader Widget
+
 ## 2.2.4
+
 - Added Skeleton Loader Widget to show a shimmer when data is loading
 - Added skeletonData on viewmodels for busy data values
+
 ## 2.2.3
 
 - Added ability to pass parameter to factories